Policy for data validation worker running on Fargate (#96)
For VPC deployments similarly to the feature server policy we want to
have a separate policy for data validation workers that also will be
hosted on Fargate.

This policy is created when the flag
`data_validation_on_fargate_enabled` is set to `true`. However, in order
to use Fargate we also need "pod_execution_role" to be created. Hence
`data_validation_on_fargate_enabled` should be switched on together with
`fargate_enabled`.

---------

Co-authored-by: Scott Ames <scott@tecton.ai>