[vpc] Add support for rift compute on VPC clusters (#185)
Changes to allow Rift compute to run on deployments set up with the
`vpc` model:

Adding an `enable_rift` boolean variable to `roles` module as well as
`emr_sample`. If enabled, a different policy file with additional
policies for rift (`eks_policy_with_rift.json`) is used. In addition, an
'offline store reader role' is created that can be assumed by the EKS
worker and has access to read offline store paths. The following
variable inputs are needed with this flag:
* `rift_compute_manager_arn` -- EKS worker role will be allowed to
AssumeRole into this role (created separately in `rift_compute` module
and passed in as input).
* `rift_ecr_repository_arn` -- EKS worker will be allowed to push to
this repo (created separately in `rift_compute` module and passed in as
input).

There are also optional parameters:
* `offline_store_key_prefix`: For cases where the offline store does not
have default `offline-store/` prefix.
* `offline_store_cmk_arns`: List of ARNs for KMS keys used to encrypt
offline store data.