feat: Set default tecton_outputs location to presigned url. (#224)
1. Set default `type` for output location config to
`tecton_hosted_presigned`, and update all READMEs / sample snippets for
module invocations to show that the URL is an expected input. The
default behavior will be to write outputs to an s3 location with URL
provided by Tecton, instead of creating a new bucket or using the
offline-store (although those remain as options).
2. Update `location_config` variable name to `outputs_location_config`
for clarity.
3. Fix: Set correct type for `vm_workload_subnet_ids` variable in
`tecton_outputs` module (`string` instead of `list(string)`).


### Testing
Tested with invocation of `dataplane_rift` module against this branch
using `?ref=tecton-outputs-update` , with a presigned URL input.
Terraform Apply worked in general + outputs written as expected.